refactor apply cn and install uninstalled packages
