+++ package.json
@@ -12,11 +12,13 @@
     "@heroicons/react": "^2.1.5",
     "@material-tailwind/react": "^2.1.10",
     "@radix-ui/react-select": "^2.1.2",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "framer-motion": "^11.12.0",
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.454.0",
     "next": "14.2.1",
+    "radix-ui": "^1.0.1",
     "react": "19.0.0-rc-02c0e824-20241028",
     "react-dom": "19.0.0-rc-02c0e824-20241028",
     "recoil": "0.7.7",
