+++ src/components/ui/dialog.tsx
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { cn } from '../../lib/utils';
+import { cn } from '../util/tailwindClassUtils';
 // import { X } from "lucide-react";
 
 const Dialog = DialogPrimitive.Root;
