+++ src/components/ui/select.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import * as SelectPrimitive from '@radix-ui/react-select';
 import { Check, ChevronDown, ChevronUp } from 'lucide-react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/components/util/tailwindClassUtils';
 
 const Select = SelectPrimitive.Root;
